+++ RyuSocks/SocksClient.cs
@@ -78,6 +78,11 @@ namespace RyuSocks
         /// <inheritdoc cref="Socket.Dispose"/>
         public void Dispose()
         {
+            if (Command is IDisposable disposableCommand)
+            {
+                disposableCommand.Dispose();
+            }
+
             _socket.Dispose();
             GC.SuppressFinalize(this);
         }
@@ -276,7 +281,7 @@ namespace RyuSocks
             }
         }
 
-        public void Connect(IPAddress address, int port)
+        public void Connect(IPEndPoint remoteEP)
         {
             if (!Authenticated)
             {
@@ -289,7 +294,7 @@ namespace RyuSocks
             }
 
             // Create the proxy command. This sends the CommandRequest to the server.
-            Command = RequestCommand.GetClientCommand()(this, new ProxyEndpoint(new IPEndPoint(address, port)));
+            Command = RequestCommand.GetClientCommand()(this, new ProxyEndpoint(remoteEP));
 
             // Process command responses until the connection is ready to be used for data or an exception is thrown.
             CommandResponse response = new();
@@ -306,6 +311,8 @@ namespace RyuSocks
                 ProcessCommandResponse(response);
             }
         }
+
+        public void Connect(IPAddress address, int port) => Connect(new IPEndPoint(address, port));
 
         public void Connect(string host, int port)
         {
@@ -336,6 +343,17 @@ namespace RyuSocks
                 Debug.Assert(receivedBytes == response.Bytes.Length);
                 ProcessCommandResponse(response);
             }
+        }
+
+        public void Disconnect()
+        {
+            if (Command is { HandlesCommunication: true })
+            {
+                Command.Disconnect();
+            }
+
+            // TODO: Set reuseSocket to true once we can handle reconnects here
+            _socket.Disconnect(false);
         }
 
         public int Send(ReadOnlySpan<byte> buffer, SocketFlags socketFlags, out SocketError errorCode)
@@ -369,8 +387,7 @@ namespace RyuSocks
 
             if (Command is { Ready: true, HandlesCommunication: true })
             {
-                errorCode = SocketError.Success;
-                return Command.Send(sendBuffer[..bufferLength]);
+                return Command.Send(sendBuffer[..bufferLength], socketFlags, out errorCode);
             }
 
             return _socket.Send(sendBuffer[..bufferLength], socketFlags, out errorCode);
@@ -390,13 +407,16 @@ namespace RyuSocks
 
             if (Command is { Ready: true, HandlesCommunication: true })
             {
-                // TODO: Fix the signature of Send/Receive methods for commands
-                bytesReceived = Command.Receive(buffer);
-                errorCode = SocketError.Success;
+                bytesReceived = Command.Receive(buffer, socketFlags, out errorCode);
             }
             else
             {
                 bytesReceived = _socket.Receive(buffer, socketFlags, out errorCode);
+            }
+
+            if (errorCode != SocketError.Success)
+            {
+                return bytesReceived;
             }
 
             if (Authenticated)
@@ -452,7 +472,7 @@ namespace RyuSocks
 
             if (Command is { Ready: true })
             {
-                return Command.SendTo(sendBuffer[..bufferLength], remoteEP);
+                return Command.SendTo(sendBuffer[..bufferLength], socketFlags, remoteEP);
             }
 
             return _socket.SendTo(sendBuffer[..bufferLength], socketFlags, remoteEP);
@@ -467,8 +487,7 @@ namespace RyuSocks
                 throw new InvalidOperationException($"{nameof(ReceiveFrom)} can only be used when receiving datagrams.");
             }
 
-            // TODO: Fix the signature of SendTo/ReceiveTo methods for commands
-            int bytesReceived = Command.ReceiveFrom(buffer, ref remoteEP);
+            int bytesReceived = Command.ReceiveFrom(buffer, socketFlags, ref remoteEP);
 
             if (Authenticated)
             {